Fixed some bugs with the save button
