+++ src/App.css
@@ -144,7 +144,7 @@ img {
   flex-direction: row;
   align-items: center;
 
-  gap: 10;
+  gap: 10px;
 }
 
 .locationText {
